fix: (#1119) prune_history will delete blocks in the range, with gaps.

+++ zilliqa/src/consensus.rs
@@ -3048,17 +3048,17 @@ impl Consensus {
     /// Removes older blocks than the specified block number, and their associated transactions.
     fn prune_history(&mut self, number: u64) -> Result<()> {
         if self.prune_interval != u64::MAX {
+            let range = self.db.available_range()?;
             let prune_at = number.saturating_sub(self.prune_interval);
-            // this may take a while, the first time pruning is activated
-            for n in (0..prune_at).rev() {
-                let block = self.db.get_canonical_block_by_number(n)?;
-                if let Some(block) = block {
-                    self.db
-                        .remove_transactions_executed_in_block(&block.hash())?;
-                    self.db.remove_block(&block)?;
-                } else {
-                    // done pruning
-                    break;
+            if range.contains(&prune_at) {
+                // this may take a while, the first time pruning is activated
+                for n in (*range.start()..prune_at).rev() {
+                    let block: Option<Block> = self.db.get_canonical_block_by_number(n)?;
+                    if let Some(block) = block {
+                        self.db
+                            .remove_transactions_executed_in_block(&block.hash())?;
+                        self.db.remove_block(&block)?;
+                    }
                 }
             }
         }
